### 2025-02-10 #7766 @kthyng
👋 @kshitijrajsharma - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. Then use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.
It looks like the problem is that your .bib file is empty. Do you mean you don't have anything to reference?
