+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: "VirtuGhan:A Virtual Computation Cube for EO Data Using On-the-Fly Tiling Computation"
+title: "VirtuGhan: A Virtual Computation Cube for EO Data Using On-the-Fly Tiling Computation"
 tags:
   - geospatial
   - data cube
@@ -19,90 +19,41 @@ bibliography: paper.bib
 ---
 
 ## Summary
-We introduce virtughan, a Python-based geospatial data pipeline designed for dynamic, on-the-fly computations on raster tiles. This system provides money-efficient and scalable tile processing by leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints, and delivering immediate results at varying zoom levels and time dimensions. Specifically, the pipeline can dynamically scale from single pixels to multiple zoom levels, applying user-defined computations or filtering steps while creating each tile. By focusing on tile-based requests, virtughan minimizes data transfer, ensures computational efficiency, and supports near-real-time data exploration to cloud-based geospatial engines, but on minimal hardware.
 
-## 1. Introduction
-Traditionally, geospatial pipelines store pre-processed raster files, which can limit flexibility when new analyses are required. virtughan breaks this limitation by computing results on the fly during tile requests. Its design is influenced by “virtual data cube” concepts and aims to provide capabilities similar to platforms like Google Earth Engine while keeping costs manageable. The system:
-- Fetches only the required data for (x, y, zoom) tiles across time.  
-- Scales single pixels to multiple zoom levels, ensuring consistent spatial footprints for advanced analyses.  
-- Integrates with Sentinel-2 STAC catalogs and COGs for partial reads, thus reducing overall I/O.  
-- Applies real-time filtering and computation formulas to select only those tiles that fully cover the area of interest.
+We introduce **VirtuGhan**, a Python-based geospatial data pipeline designed for on-the-fly computations on raster tiles. By leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints [@stacspec], VirtuGhan enables real-time data processing at multiple zoom levels and time dimensions. VirtuGhan focses on on-demand tile computation, computes results on demand, minimizing data transfers and infrastructure overhead [@rio-tiler; @mercantile] for entire scene. This approach is cost-effective because only the necessary tiles (i.e., bounding boxes of interest) are read & computation is applied on tiles when fetching from source imagery [@sentinel2cogs]. The framework supports user-defined band math, multi-temporal analyses, partial reads from Cloud Optimized Sentinel-2 data, and a caching mechanism for repeat requests [@fastapi]. Ultimately, **VirtuGhan** provides a scalable, open-source platform for modern geospatial data analytics, running efficiently even on minimal hardware.
 
-## 2. Background
-**Tile-Based Delivery**: Common for web mapping platforms, tiling offers a straightforward approach to request a small portion of data for a given bounding box at a particular zoom.  
+## Statement of Need
 
-**Data Cubes**: Provide a conceptual framework to store spatiotemporal data, but can be large. virtughan uses a “virtual” approach, reading only the data it needs without storing.  
+Large-scale Earth Observation (EO) analytics often demand substantial storage and processing power, especially when frequently querying large satellite datasets at varied resolutions. Traditional pipelines that store entire scenes or “data cubes” comes with disadvantage of  cost it takes to maintain. **VirtuGhan** addresses this gap by:
 
-**STAC and COGs**: Allow the system to quickly discover relevant imagery and perform efficient windowed reads on cloud-hosted GeoTIFFs.  
[…]
