+++ paper.bib
@@ -1,1 +1,53 @@
+@misc{rio-tiler,
+  title        = {rio-tiler: Rasterio plugin for reading and creating mosaics},
+  howpublished = {\url{https://github.com/cogeotiff/rio-tiler}},
+  note         = {Accessed: 2025-02-08}
+}
 
+@misc{mercantile,
+  title        = {mercantile: A Python library for working with map tiles},
+  howpublished = {\url{https://github.com/mapbox/mercantile}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{stacspec,
+  title        = {SpatioTemporal Asset Catalog (STAC) Specification},
+  howpublished = {\url{https://stacspec.org}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{fastapi,
+  title        = {FastAPI: A modern, fast web framework for Python},
+  howpublished = {\url{https://fastapi.tiangolo.com}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{rasterio,
+  title        = {Rasterio Documentation},
+  howpublished = {\url{https://rasterio.readthedocs.io}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{sentinel2cogs,
+  title        = {Sentinel-2 L2A COGs on AWS},
+  howpublished = {\url{https://registry.opendata.aws/sentinel-2-l2a-cogs/}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{cog,
+  title        = {Cloud Optimized GeoTIFF (COG) specification},
+  howpublished = {\url{https://www.cogeo.org}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{earthsearch,
+  title        = {earth-search: Earth-Search STAC API for Landsat and Sentinel-2 COGs},
+  howpublished = {\url{https://github.com/Element84/earth-search}},
+  note         = {Accessed: 2025-02-08}
+}
+
+@misc{gee,
+  title        = {Google Earth Engine},
+  howpublished = {\url{https://earthengine.google.com}},
+  note         = {Accessed: 2025-02-08}
+}
