+++ paper.md
@@ -48,7 +48,7 @@ The processed tiles (e.g., PNG or JPEG) can be cached. If an identical tile requ
 
 # Figures
 
-![Flowchart-virtughan](https://github.com/user-attachments/assets/cc6a67af-01ca-40d6-8442-1a32b5593e2e)
+![Flowchart-virtughan](flowchart-virtughan.png)
 
 ## Future Directions
 
