### 2025-02-11 #7766 @kshitijrajsharma
> 👋 [@kshitijrajsharma](https://github.com/kshitijrajsharma) - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. Then use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment. It looks like the problem is that your .bib file is empty. Do you mean you don't have anything to reference?

Dear @kthyng  I have added references and pictures , I also have github action to check  pdf generation : https://github.com/kshitijrajsharma/virtughan/blob/master/.github/workflows/draft-pdf.yml

Which is passing : https://github.com/kshitijrajsharma/virtughan/actions/runs/13260909671/job/37017089767

But here it fails , it is possible editorialbot is  checking out the version itself not the latest changes ? How do I make sure it takes the latest changes ?
