### 2025-02-11 #7766 @kthyng
@kshitijrajsharma It is using the master branch. I am not sure why it isn't working. Maybe the old paper.bib is cached somehow from before you made changes to it. @openjournals/dev Could you help figure out why the paper building doesn't seem to be accounting for recent changes to the bib file?
