### 2025-02-11 #7766 @kthyng
Ah @kshitijrajsharma I see a paper.md in the main directory and in `docs`. Which do you intend to use? Maybe it is finding the one in docs and expected the bib file to be with it.
