### 2025-02-11 c47573e
Delete docs/paper.md
