+++ docs/paper.md
@@ -1,108 +1,0 @@
----
-title: "VirtuGhan:A Virtual Computation Cube for EO Data Using On-the-Fly Tiling Computation"
-tags:
-  - geospatial
-  - data cube
-  - STAC
-  - COG
-  - Python
-authors:
-  - name: "Kshitij Raj Sharma"
-    affiliation: 1
-  - name: "Upendra Oli"
-    affiliation: 1
-affiliations:
-  - name: "Paris Lodron University of Salzburg, Austria"
-    index: 1
-date: "2025-02-08"
-bibliography: paper.bib
----
-
-## Summary
-We introduce virtughan, a Python-based geospatial data pipeline designed for dynamic, on-the-fly computations on raster tiles. This system provides money-efficient and scalable tile processing by leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints, and delivering immediate results at varying zoom levels and time dimensions. Specifically, the pipeline can dynamically scale from single pixels to multiple zoom levels, applying user-defined computations or filtering steps while creating each tile. By focusing on tile-based requests, virtughan minimizes data transfer, ensures computational efficiency, and supports near-real-time data exploration to cloud-based geospatial engines, but on minimal hardware.
-
-## 1. Introduction
-Traditionally, geospatial pipelines store pre-processed raster files, which can limit flexibility when new analyses are required. virtughan breaks this limitation by computing results on the fly during tile requests. Its design is influenced by “virtual data cube” concepts and aims to provide capabilities similar to platforms like Google Earth Engine while keeping costs manageable. The system:
-- Fetches only the required data for (x, y, zoom) tiles across time.  
-- Scales single pixels to multiple zoom levels, ensuring consistent spatial footprints for advanced analyses.  
-- Integrates with Sentinel-2 STAC catalogs and COGs for partial reads, thus reducing overall I/O.  
-- Applies real-time filtering and computation formulas to select only those tiles that fully cover the area of interest.
-
-## 2. Background
-**Tile-Based Delivery**: Common for web mapping platforms, tiling offers a straightforward approach to request a small portion of data for a given bounding box at a particular zoom.  
-
-**Data Cubes**: Provide a conceptual framework to store spatiotemporal data, but can be large. virtughan uses a “virtual” approach, reading only the data it needs without storing.  
-
-**STAC and COGs**: Allow the system to quickly discover relevant imagery and perform efficient windowed reads on cloud-hosted GeoTIFFs.  
-
-**FastAPI**: Powers asynchronous requests, making it a suitable technology for scalable microservices serving tiles.
-
-## 3. System Overview and Methodology
-
-### 3.1 Virtual Data Cube Concept
-At the core is a virtual data cube:
-- No massive pre-stacked arrays: Instead, STAC items indicate where each requested tile’s data can be fetched.  
-- Band Math & Corrections: The system can apply formulas or filtering steps (e.g., remove cloud-covered or incomplete scenes) at tile creation time.  
-- Temporal Dimension: Maintains the same (x, y) footprint across time, letting users visualize changes or compute multi-temporal metrics without re-downloading entire scenes.
-
-### 3.2 Data Flow
-1. **Tile Request**: Front-end or an API user requests (z, x, y)—plus time, formula, or filter criteria.  
-2. **Bounding Box Calculation**: mercantile determines the bounding box for the tile.  
-3. **STAC Query & Filtering**: The pipeline searches STAC for relevant Sentinel-2 scenes that fully cover the bounding box and time range.  
-4. **Windowed Read & Original Resolution**: Partial reads from COGs at the original or resampled resolution.  
[…]
