### 2025-02-11 #7766 @kthyng
@kshitijrajsharma I am going to ping the editorial board to make sure your submission is in scope for JOSS. This will take a few weeks, thanks for your patience.
