### 2025-02-11 #7766 @kshitijrajsharma
> @kshitijrajsharma I am going to ping the editorial board to make sure your submission is in scope for JOSS. This will take a few weeks, thanks for your patience.

Thank you for your effort , I will also go through the list you provided and make sure we check all
