+++ paper.md
@@ -20,16 +20,14 @@ bibliography: paper.bib
 
 # Summary
 
-We introduce **VirtuGhan**, a Python-based geospatial data pipeline designed for on-the-fly computations on raster tiles. By leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints [@stacspec], VirtuGhan enables real-time data processing at multiple zoom levels and time dimensions. VirtuGhan focses on on-demand tile computation, computes results on demand, minimizing data transfers and infrastructure overhead [@rio-tiler; @mercantile] for entire scene. This approach is cost-effective because only the necessary tiles (i.e., bounding boxes of interest) are read & computation is applied on tiles when fetching from source imagery [@sentinel2cogs]. The framework supports user-defined band math, multi-temporal analyses, partial reads from Cloud Optimized Sentinel-2 data, and a caching mechanism for repeat requests [@fastapi]. Ultimately, **VirtuGhan** provides a scalable, open-source platform for modern geospatial data analytics, running efficiently even on minimal hardware.
+**VirtuGhan** is a Python-based geospatial data pipeline designed for on-the-fly computation on raster tiles, enabling real-time processing of Earth Observation (EO) data. By leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints, it efficiently processes data at multiple zoom levels and time dimensions without requiring full-scene precomputations.
+Unlike traditional processing pipelines that rely on pre-aggregated layers, VirtuGhan retrieves and processes only the necessary tiles on demand, significantly reducing storage needs and computational overhead. This efficient approach minimizes data transfer and optimizes infrastructure utilization, making it a cost-effective solution for cloud-based EO analytics [@rio-tiler; @mercantile; @sentinel2cogs]. The framework supports user-defined band mathematics, multi-temporal analysis, and partial reads from Sentinel-2 data, with an integrated caching mechanism for optimizing repeated queries [@fastapi].
+As an open-source and scalable platform, VirtuGhan provides a lightweight alternative to traditional data cube architectures. Its optimized tile-based computation allows it to run efficiently even on minimal hardware, making large-scale satellite imagery processing more accessible for researchers, analysts, and developers.
 
 # Statement of Need
 
-Large-scale Earth Observation (EO) analytics often demand substantial storage and processing power, especially when frequently querying large satellite datasets at varied resolutions. Traditional pipelines that store entire scenes or “data cubes” comes with disadvantage of  cost it takes to maintain. **VirtuGhan** addresses this gap by:
-
-- Retrieving and processing only the tiles required, reducing bandwidth, storage, and memory usage.
-- Applying the computation while fetching the tile for a same area of interest in multiple hierarchy of tiles
-- Providing a virtual “data cube” model in which pixel transformations are computed on-the-fly, negating the need to store massive multi-band, multi-temporal archives.  
-- Facilitating google earth engine alike computation, interactive geospatial exploration for scientists, engineers, and data analysts who need to quickly visualize or process recent satellite data without incurring large cloud-computing costs in different zoom levels of same area.
[…]
