+++ paper.md
@@ -47,7 +47,7 @@ The processed tiles (e.g., PNG or JPEG) can be cached. If an identical tile requ
 
 # Figures
 
-![Flowchart-virtughan](flowchart-virtughan1.jpg)
+![Flowchart-virtughan](flowchart-virtughan2.jpg)
 
 ## Future Directions
 
