### 2025-02-16 48b121a
paper: updated acknowledgement
