+++ paper.md
@@ -64,6 +64,6 @@ VirtuGhan enables real-time geospatial data processing for various Earth observa
 
 # Acknowledgments
 
-We extend our gratitude to the maintainers of foundational GIS libraries such as Rasterio [@rasterio], rio-tiler [@rio-tiler], Mercantile [@mercantile], and FastAPI [@fastapi]. We also thank colleagues and mentors at Paris Lodron University of Salzburg and in the Copernicus Masters in Digital Earth program for their insights and support.
+We extend our gratitude to the maintainers of foundational GIS libraries such as Rasterio [@rasterio], rio-tiler [@rio-tiler], Mercantile [@mercantile], and FastAPI [@fastapi]. We extend our gratitude to our colleagues and mentors at the Copernicus Masters in Digital Earth program, co-funded by the European Union and hosted at Paris Lodron University of Salzburg, for their valuable insights and support. A special thanks to Associate Professor Dirk Tiede and Dr. Martin Sudmanns for their guidance on existing data cube methodologies.
 
 # References
