+++ paper.md
@@ -20,9 +20,9 @@ bibliography: paper.bib
 
 # Summary
 
-**VirtuGhan** is a Python-based geospatial data pipeline designed for on-the-fly computation on raster tiles, enabling real-time processing of Earth Observation (EO) data. By leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints, it efficiently processes data at multiple zoom levels and time dimensions without requiring full-scene precomputations.
-Unlike traditional processing pipelines that rely on pre-aggregated layers, VirtuGhan retrieves and processes only the necessary tiles on demand, significantly reducing storage needs and computational overhead. This efficient approach minimizes data transfer and optimizes infrastructure utilization, making it a cost-effective solution for cloud-based EO analytics [@rio-tiler; @mercantile; @sentinel2cogs]. The framework supports user-defined band mathematics, multi-temporal analysis, and partial reads from Sentinel-2 data, with an integrated caching mechanism for optimizing repeated queries [@fastapi].
-As an open-source and scalable platform, VirtuGhan provides a lightweight alternative to traditional data cube architectures. Its optimized tile-based computation allows it to run efficiently even on minimal hardware, making large-scale satellite imagery processing more accessible for researchers, analysts, and developers.
+**VirtuGhan** is a Python-based geospatial data pipeline designed for on-the-fly computations on raster tiles. By leveraging Cloud-Optimized GeoTIFFs (COGs) and SpatioTemporal Asset Catalog (STAC) endpoints [@stacspec], VirtuGhan enables real-time data processing across multiple zoom levels and time dimensions.  
+VirtuGhan focuses on on-demand tile computation, dynamically computing results as needed while minimizing data transfers and infrastructure overhead for entire scenes [@rio-tiler; @mercantile]. This cost-effective approach ensures that only the necessary tiles (i.e., bounding boxes of interest) are retrieved, with computations applied directly during data access [@sentinel2cogs]. The framework supports user-defined band math, multi-temporal analyses, and partial reads from Cloud-Optimized Sentinel-2 data, along with a caching mechanism to optimize repeated requests [@fastapi].  
+Ultimately, VirtuGhan provides a scalable, open-source platform for modern geospatial data analytics, offering efficient, and real-time processing. Its optimized tile-based computation allows it to run efficiently even on minimal hardware, making large-scale satellite imagery processing more accessible for researchers, analysts, and developers.
 
 # Statement of Need
 
