### 2025-02-16 4064da8
paper: application areas update2
