### 2025-02-16 23c2ed4
taking applicaiton areas above future directions
