### 2025-02-19 e563504
paper: updated citation format in bib
