+++ paper.bib
@@ -51,3 +51,15 @@
   howpublished = {\url{https://earthengine.google.com}},
   note         = {Accessed: 2025-02-08}
 }
+
+@misc{sudmanns,
+  title        = {Sudmanns et al., 2019},
+  howpublished = {\url{https://www.tandfonline.com/doi/full/10.1080/17538947.2019.1585976}},
+  note         = {Accessed: 2025-02-10}
+}
+
+@misc{giuliani,
+  title        = {Giuliani & Camara, 2019},
+  howpublished = {\url{https://www.mdpi.com/2306-5729/4/4/147}},
+  note         = {Accessed: 2025-02-10}
+}
