### 2025-02-20 f579601
warning messages for bigger areas than 500 sq km | and point draw with bounding box of 200 sq km.
