### 2025-02-20 c571ea2
draw point bbox style
