### 2025-02-20 732f49c
warning for AOI - for map window bbox
