### 2025-02-20 a39670d
reduce radius for point AOI
