### 2025-02-20 2a0101d
fix: show warning on window showing even for draw
