+++ static/js/export_download.js
@@ -1,4 +1,5 @@
 var completed_log = false;
+downloading = false;
       document.getElementById("export-map-view-button").addEventListener('click', function() {
         completed_log = false;
         var analyzeChecked = document.getElementById("analyze-data").checked;
@@ -39,19 +40,23 @@ var completed_log = false;
         }
 
         //show warning if the area of bbox is greater than 500 SQ Km.
-        var bbox_rectangle = L.rectangle(bounds, {fillOpacity: 0.1, opacity: 0.6});
-        var geojson = bbox_rectangle.toGeoJSON();
-
-        // Calculate the area using turf.js
-        var area = turf.area(geojson);
-
-        // Convert area to square kilometers (optional)
-        var areaKm2 = area / 1000000;
-        // console.log('Area: ' + areaKm2.toFixed(2) + ' square kilometers');
-
-        if(areaKm2 > 500){
-          // showMessage('success', "message");
-          showMessage('warning', "Zoom in and reduce the size of your area of interest. <br/> Eg. smaller AOI than 500 SQ.Km. Sorry, this is due to limited server specs.");
+        var aoiText = document.getElementById("selected-filter-value-bbox").innerText;
+        // console.log(aoiText);
+        if(aoiText == "Map Window"){
+          var bbox_rectangle = L.rectangle(bounds, {fillOpacity: 0.1, opacity: 0.6});
+          var geojson = bbox_rectangle.toGeoJSON();
+  
+          // Calculate the area using turf.js
+          var area = turf.area(geojson);
+  
+          // Convert area to square kilometers (optional)
+          var areaKm2 = area / 1000000;
+          // console.log('Area: ' + areaKm2.toFixed(2) + ' square kilometers');
+  
+          if(areaKm2 > 500){
+            // showMessage('success', "message");
+            showMessage('warning', "Zoom in and reduce the size of your area of interest. <br/> Eg. smaller AOI than 500 SQ.Km. Sorry, this is due to limited server specs.");
+          }
         }
         //show warning end.
         
@@ -240,10 +245,13 @@ var completed_log = false;
         if (computeLayer) {
             map.removeLayer(computeLayer);
         }
+        downloading = true; // for loader to load until it downloads file
+
         fetch(tifUrl)
           .then(response => response.arrayBuffer())
           .then(arrayBuffer => {
             parseGeoraster(arrayBuffer).then(georaster => {
+              downloading = false; //remove loader afer file is downloaded.
               console.log("georaster:", georaster);
               geoRaster = georaster;
 
@@ -291,7 +299,7 @@ var completed_log = false;
                        // Get continuous color
                   }
               });
-              showLoaderOnMap(computeLayer);
+              showLoaderOnMap(computeLayer, false);
               map.addLayer(computeLayer);
               updateLegend(selectedPalette);
 
