### 2025-02-20 d28d5c2
loading while layer is being fetched for visualization
