+++ static/js/map.js
@@ -153,23 +153,29 @@ var map = L.map("map").setView([28.202082, 83.987222], 10);
         loaderControl.addTo(map);
 
       function showLoaderOnMap(layer, downloading){
-        console.log("entered");
-        // Show loader when tiles start loading
-        layer.on('loading', () => {
-          loaderControl.getContainer().style.display = 'block';
-        });
+        // console.log("entered");
+        if(layer){
+          // Show loader when tiles start loading
+          layer.on('loading', () => {
+            loaderControl.getContainer().style.display = 'block';
+          });
 
-        // Hide loader when tiles are fully loaded
-        layer.on('load', () => {
+          // Hide loader when tiles are fully loaded
+          layer.on('load', () => {
+            loaderControl.getContainer().style.display = 'none';
+          });
           loaderControl.getContainer().style.display = 'none';
-        });
-
-        if(downloading){
-          loaderControl.getContainer().style.display = 'block';
         }
         else{
-          loaderControl.getContainer().style.display = 'none';
+          if(downloading){
+            loaderControl.getContainer().style.display = 'block';
+          }
+          else{
+            loaderControl.getContainer().style.display = 'none';
+          }
+          
         }
+        
       }
 
       var geojsonLayer;
