### 2025-02-20 8b95d18
remove loader if there is error
