### 2025-02-20 4876dd6
message type error: dont hide automatically
