### 2025-02-20 7bcc1b1
error and warning texts in a message
