+++ README.md
@@ -177,8 +177,8 @@ If you love what we’re building, consider buying us a coffee ☕ to keep the p
 
 ## Acknowledgment
 
-This project was undertaken as part of the project work for our master's program , Coopernicus Masters in Digital Earth. 
-
+This project was initiated during the project work of our master's program , Coopernicus Masters in Digital Earth. 
+We are thankful to all those involved and supported us from the program. 
 <p align="left">
   <img src="https://github.com/user-attachments/assets/2f0555f8-67c3-49da-a0e8-037bdfd4ce10" alt="CMIDE-InLine-logoCMYK" style="width:200px;"/>
   <img src="https://github.com/user-attachments/assets/e553c675-f8e5-440a-b50f-625d0ce4f0c9" alt="EU_POS_transparent" style="width:200px;"/>
