### 2025-02-21 82de950
remove outdated image reference from README.md
