### 2025-02-21 92d6130
docs(mkdocs): adds mkdocs setup
4
