+++ src/vcube/utils.py
@@ -11,6 +11,18 @@ STAC_API_URL = "https://earth-search.aws.element84.com/v1/search"
 
 
 def search_stac_api(bbox, start_date, end_date, cloud_cover):
+    """
+    Search the STAC API for satellite images.
+
+    Parameters:
+    bbox (list): Bounding box coordinates [min_lon, min_lat, max_lon, max_lat].
+    start_date (str): Start date for the search (YYYY-MM-DD).
+    end_date (str): End date for the search (YYYY-MM-DD).
+    cloud_cover (int): Maximum allowed cloud cover percentage.
+
+    Returns:
+    list: List of features found in the search.
+    """
     search_params = {
         "collections": ["sentinel-2-l2a"],
         "datetime": f"{start_date}T00:00:00Z/{end_date}T23:59:59Z",
@@ -42,6 +54,18 @@ def search_stac_api(bbox, start_date, end_date, cloud_cover):
 
 
 async def search_stac_api_async(bbox_geojson, start_date, end_date, cloud_cover):
+    """
+    Asynchronously search the STAC API for satellite images.
+
+    Parameters:
+    bbox_geojson (dict): Bounding box in GeoJSON format.
+    start_date (str): Start date for the search (YYYY-MM-DD).
+    end_date (str): End date for the search (YYYY-MM-DD).
+    cloud_cover (int): Maximum allowed cloud cover percentage.
+
+    Returns:
+    list: List of features found in the search.
+    """
     search_params = {
         "collections": ["sentinel-2-l2a"],
         "datetime": f"{start_date}T00:00:00Z/{end_date}T23:59:59Z",
@@ -75,6 +99,13 @@ async def search_stac_api_async(bbox_geojson, start_date, end_date, cloud_cover)
 
 
 def zip_files(file_list, zip_path):
+    """
+    Zip a list of files.
+
+    Parameters:
+    file_list (list): List of file paths to zip.
+    zip_path (str): Path to the output zip file.
+    """
     with zipfile.ZipFile(zip_path, "w", compression=zipfile.ZIP_DEFLATED) as zipf:
         for file in file_list:
             zipf.write(file, os.path.basename(file))
@@ -84,6 +115,15 @@ def zip_files(file_list, zip_path):
 
 
 def filter_latest_image_per_grid(features):
+    """
+    Filter the latest image per grid.
+
+    Parameters:
+    features (list): List of features to filter.
+
+    Returns:
+    list: List of filtered features.
+    """
     grid_latest = {}
     for feature in features:
         grid = feature["id"].split("_")[1]
@@ -97,6 +137,16 @@ def filter_latest_image_per_grid(features):
 
 
 def filter_intersected_features(features, bbox):
+    """
+    Filter features that intersect with the bounding box.
+
+    Parameters:
+    features (list): List of features to filter.
+    bbox (list): Bounding box coordinates [min_lon, min_lat, max_lon, max_lat].
+
+    Returns:
+    list: List of filtered features.
+    """
     bbox_polygon = box(bbox[0], bbox[1], bbox[2], bbox[3])
     return [
         feature
@@ -106,6 +156,15 @@ def filter_intersected_features(features, bbox):
 
 
 def remove_overlapping_sentinel2_tiles(features):
+    """
+    Remove overlapping Sentinel-2 tiles.
+
+    Parameters:
+    features (list): List of features to process.
+
+    Returns:
+    list: List of non-overlapping features.
+    """
     if not features:
         return []
 
@@ -132,6 +191,16 @@ def remove_overlapping_sentinel2_tiles(features):
 
 
 def aggregate_time_series(data, operation):
+    """
+    Aggregate a time series of data.
+
+    Parameters:
+    data (list): List of data arrays to aggregate.
[…]

[…]
