### 2025-02-21 0a9e788
ci: update MkDocs workflow to trigger on release events
