+++ .github/workflows/mkdocs.yml
@@ -1,12 +1,9 @@
 name: Deploy MkDocs
 
 on:
-  push:
-    branches:
-      - master
-  pull_request:
-    branches:
-      - master
+  release:
+    types: [published]
+  workflow_dispatch:
 
 permissions:
   contents: write
