### 2025-02-21 f4ef2a7
docs: update mkdocs configuration and add API reference documentation
