+++ docs/src/engine.md
@@ -1,0 +1,3 @@
+# Engine Module
+
+::: vcube.engine
