+++ docs/src/tile.md
@@ -1,0 +1,3 @@
+# Tile Module
+
+::: vcube.tile
