+++ docs/src/extract.md
@@ -1,0 +1,3 @@
+# Extract Module
+
+::: vcube.extract
