+++ docs/src/utis.md
@@ -1,0 +1,3 @@
+# Utils Module
+
+::: vcube.utils
