### 2025-02-21 38475b0
docs: restructure Examples section in mkdocs configuration
