### 2025-02-21 3d43bb7
docs: update README with new notebook link and add playground notebook
