### 2025-02-25 d37e67e
updated reference citations
