+++ paper.bib
@@ -52,14 +52,28 @@
   note         = {Accessed: 2025-02-08}
 }
 
-@misc{sudmanns,
-  title        = {Sudmanns et al., 2019},
-  howpublished = {\url{https://www.tandfonline.com/doi/full/10.1080/17538947.2019.1585976}},
-  note         = {Accessed: 2025-02-10}
+
+@article{Sudmanns:2019,
+    url = {https://www.tandfonline.com/doi/full/10.1080/17538947.2019.1585976},
+    Archiveprefix = {DOI},
+    Author = {{Sudmanns}, M. and {Tiede}, D. and {Lang}, S. and {Bergstedt}, H. and {Trost}, G. and {Augustin}, H. and {Baraldi}, A. and {Blaschke}, T.},
+    Eprint = {10.1080/17538947.2019.1585976},
+    Journal = {International Journal of Digital Earth},
+    Keywords = {Digital Earth, data access, satellite data portals, object-based image analysis (OBIA), remote sensing workflow},
+    Month = {Mar},
+    Title = {{Big Earth data: disruptive changes in Earth observation data management and analysis?}},
+    Year = {2019}
 }
 
-@misc{giuliani,
-  title        = {Giuliani & Camara, 2019},
-  howpublished = {\url{https://www.mdpi.com/2306-5729/4/4/147}},
-  note         = {Accessed: 2025-02-10}
+@article{Giuliani:2019,
+    url = {https://www.mdpi.com/2306-5729/4/4/147},
+    Archiveprefix = {DOI},
+    Author = {{Giuliani}, G. and {Camara}, G. and {Killough}, B. and {Minchin}, S.},
+    Eprint = {10.3390/data4040147},
+    Journal = {Data},
+    Keywords = {open science, reproducibility, earth observations, data cube, analysis ready data, remote sensing, satellite imagery},
+    Month = {Nov},
+    Title = {{Earth Observation Open Science: Enhancing Reproducible Science Using Data Cubes}},
+    Year = {2019}
 }
+
