### 2025-04-06 16d8bce
feat: parallel processing: added uid on export and log
