+++ static/js/export_download.js
@@ -112,6 +112,12 @@ downloading = false;
         .then(response => response.json()) 
         .then(data => {
           console.log('Success:', data);
+
+          // Store the UID in localStorage
+          if (data.uid) {
+            localStorage.setItem('UID', data.uid);
+          }
+
         })
         .catch((error) => {
           console.error('Error:', error);
@@ -127,7 +133,12 @@ downloading = false;
         //show the progress to the user
         
         function checkProcessingStatus() {
-          fetch('/logs', { 
+          const uid = localStorage.getItem('UID');
+          console.log("UID: ", uid)
+
+          const logUrl = "/logs?uid="+uid;
+          
+          fetch(logUrl, { 
             method: 'GET' 
           })
             .then(response => response.text())
