+++ static/js/export_download.js
@@ -134,10 +134,10 @@ downloading = false;
         
         function checkProcessingStatus() {
           const uid = localStorage.getItem('UID');
-          console.log("UID: ", uid)
+          // console.log("UID: ", uid)
 
           const logUrl = "/logs?uid="+uid;
-          
+
           fetch(logUrl, { 
             method: 'GET' 
           })
@@ -155,7 +155,7 @@ downloading = false;
                     if(analyzeChecked){
                       document.getElementById("layerSwitcherBox").classList.remove("hidden");
                       if(document.getElementById("operation").checked){
-                        plotGeoTIFF('static/export/custom_band_output_aggregate.tif');
+                        plotGeoTIFF('static/export/'+uid+'/custom_band_output_aggregate.tif');
                         completed_log = true;
                       }
                     }
