### 2025-04-06 cf1b58b
feat: enable smart filters automatically if date range is higher
