### 2025-04-06 090f2e2
feat: messages displayed for information in addition to warning and errors.
