### 2025-04-18 #8078 @editorialbot
[REVIEW]: VirtuGhan:A Virtual Computation Cube for EO Data Using On-the-Fly Tiling Computation
**Submitting author:** @kshitijrajsharma (<a href="http://orcid.org/0000-0002-2123-3917">Kshitij Raj Sharma</a>)
**Repository:** https://github.com/kshitijrajsharma/virtughan/
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.0
