### 2025-04-24 5b0fe52
fix: add validation for Sentinel-2 bands and resolution mismatch
