### 2025-04-24 1fc15d9
fix: return JSON response with status code for background processing initiation
