+++ src/vcube/extractors/sentinel2_extractor1.py
@@ -1,0 +1,224 @@
+import os
+import sys
+from concurrent.futures import ThreadPoolExecutor, as_completed
+
+import numpy as np
+import rasterio
+from rasterio.enums import Resampling
+from rasterio.warp import reproject
+from tqdm import tqdm
+
+from ..utils.common import (
+    filter_intersected_features,
+    search_stac_api,
+    smart_filter_images,
+    zip_files,
+)
+from ..utils.sentinel2_utils import remove_overlapping_sentinel2_tiles
+from .extractor_common import ExtractorCommon
+
+
+# ✅ Sentinel-2 supported bands
+VALID_BANDS = {
+    "red": "Red - 10m",
+    "green": "Green - 10m",
+    "blue": "Blue - 10m",
+    "nir": "NIR 1 - 10m",
+    "swir22": "SWIR 2.2μm - 20m",
+    "rededge2": "Red Edge 2 - 20m",
+    "rededge3": "Red Edge 3 - 20m",
+    "rededge1": "Red Edge 1 - 20m",
+    "swir16": "SWIR 1.6μm - 20m",
+    "wvp": "Water Vapour (WVP)",
+    "nir08": "NIR 2 - 20m",
+    "aot": "Aerosol optical thickness (AOT)",
+    "coastal": "Coastal - 60m",
+    "nir09": "NIR 3 - 60m",
+}
+class ExtractProcessor(ExtractorCommon):
+    """
+    Processor for extracting Sentinel-2 bands.
+    """
+
+    def __init__(
+        self,
+        bbox,
+        start_date,
+        end_date,
+        cloud_cover,
+        bands_list,
+        output_dir,
+        log_file=sys.stdout,
+        workers=1,
+        zip_output=False,
+        smart_filter=True,
+    ):
+        self.bbox = bbox
+        self.start_date = start_date
+        self.end_date = end_date
+        self.cloud_cover = cloud_cover
+        self.bands_list = bands_list
+        self.output_dir = output_dir
+        self.log_file = log_file
+        self.workers = workers
+        self.zip_output = zip_output
+        self.crs = None
+        self.transform = None
+        self.use_smart_filter = smart_filter
+
+        self._validate_bands_list()
+
+    def _validate_bands_list(self):
+        invalid_bands = [band for band in self.bands_list if band not in VALID_BANDS]
+        if invalid_bands:
+            raise ValueError(
+                f"Invalid band names: {', '.join(invalid_bands)}. "
+                f"Band names should be one of: {', '.join(VALID_BANDS.keys())}"
+            )
+
+    def _get_band_urls(self, features):
+        return [
+            [feature["assets"][band]["href"] for band in self.bands_list]
+            for feature in features
+        ]
+    def _fetch_and_save_bands(self, band_urls, feature_id):
+        try:
+            bands = []
+            bands_meta = []
+            resolutions = []
+
+            for band_url in band_urls:
+                with rasterio.open(band_url) as band_cog:
+                    resolutions.append(band_cog.res)
+
+            lowest_resolution = max(resolutions, key=lambda res: res[0] * res[1])
+
+            for band_url in band_urls:
+                with rasterio.open(band_url) as band_cog:
+                    min_x, min_y, max_x, max_y = self._transform_bbox(band_cog.crs)
+                    band_window = self._calculate_window(
+                        band_cog, min_x, min_y, max_x, max_y
+                    )
+
+                    if self._is_window_out_of_bounds(band_window):
+                        return None
+
+                    self.crs = band_cog.crs
+                    self.transform = band_cog.transform
+
+                    band_data = band_cog.read(1, window=band_window).astype(float)
+
+                    if band_cog.res != lowest_resolution:
+                        scale_factor_x = band_cog.res[0] / lowest_resolution[0]
[…]
