+++ src/vcube/extractors/extractor_common.py
@@ -1,0 +1,48 @@
+import os
+import numpy as np
+import rasterio
+from pyproj import Transformer
+from rasterio.windows import from_bounds
+
+
+class ExtractorCommon:
+    """
+    Base class for common extractor utilities.
+    """
+
+    def _transform_bbox(self, crs):
+        transformer = Transformer.from_crs("epsg:4326", crs, always_xy=True)
+        min_x, min_y = transformer.transform(self.bbox[0], self.bbox[1])
+        max_x, max_y = transformer.transform(self.bbox[2], self.bbox[3])
+        return min_x, min_y, max_x, max_y
+
+    def _calculate_window(self, cog, min_x, min_y, max_x, max_y):
+        return from_bounds(min_x, min_y, max_x, max_y, cog.transform)
+
+    def _is_window_out_of_bounds(self, window):
+        return (
+            window.col_off < 0
+            or window.row_off < 0
+            or window.width <= 0
+            or window.height <= 0
+        )
+    def _save_geotiff(self, bands, output_file, bands_meta=None):
+        band_shape = bands.shape
+        nodata_value = -9999
+        bands = np.where(np.isnan(bands), nodata_value, bands)
+        with rasterio.open(
+            output_file,
+            "w",
+            driver="GTiff",
+            height=bands.shape[1],
+            width=bands.shape[2],
+            count=len(bands),
+            dtype=bands.dtype,
+            crs=self.crs,
+            transform=self.transform,
+            nodata=nodata_value,
+        ) as dst:
+            for band in range(1, band_shape[0] + 1):
+                dst.write(bands[band - 1], band)
+                if bands_meta:
+                    dst.set_band_description(band, bands_meta[band - 1])

[…]
