### 2025-05-04 31084fa
added landsat option in live in frontend
