### 2025-05-04 #8078 @jabhalla
@rwegener2 @kshitijrajsharma

While the VirtuGhan paper is well-written and provides useful context on the challenges of Big Earth Data, the manuscript would benefit from additional detail in two areas.

First, although VirtuGhan is described as efficient and scalable, the paper does not include empirical benchmarks or quantitative results to substantiate these performance claims.

Second, while the authors discuss the limitations of traditional data cubes and cloud-based platforms, a more direct comparison with other commonly used real time geospatial data processing frameworks would help clarify VirtuGhan’s practical advantages and unique contributions.

Thanks
