+++ src/vcube/engines/landsat89_engine.py
@@ -6,10 +6,11 @@ import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 import rasterio
-from PIL import Image
 from tqdm import tqdm
-from pyproj import Transformer
-from rasterio.windows import from_bounds
+from PIL import Image
+from planetary_computer import sign
+from urllib.parse import urlparse
+from datetime import datetime
 
 from ..utils.common import (
     filter_intersected_features,
@@ -20,14 +21,6 @@ from ..utils.common import (
 from ..utils.landsat89_utils import (
     remove_overlapping_landsat_tiles,
 )
-# from ..engines.engine_common import (
-#     _transform_bbox as transform_bbox,
-#     calculate_window,
-#     is_window_out_of_bounds,
-#     pad_array,
-#     add_text_to_image,
-#     create_gif,
-# )
 from ..engines.engine_common import EngineCommon
 
 matplotlib.use("Agg")
@@ -70,58 +63,46 @@ class LandsatProcessor(EngineCommon):
         self.intermediate_images = []
         self.intermediate_images_with_text = []
         self.use_smart_filter = smart_filter
-    # def _transform_bbox(self, crs):
-    #     return transform_bbox(self.bbox, crs)
-
-    # def _calculate_window(self, cog, min_x, min_y, max_x, max_y):
-    #     return calculate_window(cog, min_x, min_y, max_x, max_y)
-
-    # def _is_window_out_of_bounds(self, window):
-    #     return is_window_out_of_bounds(window)
+    
+    def _parse_dates(self, image_name):
+        date = image_name.split("_")[3]
+        parsed_date = datetime.strptime(date, "%Y%m%d").strftime("%Y-%m-%d")
+        return parsed_date
 
     def _get_band_urls(self, features):
-        band1_urls = [feature["assets"][self.band1]["href"] for feature in features]
+        band1_urls = [sign(feature["assets"][self.band1]["href"]) for feature in features]
         band2_urls = (
-            [feature["assets"][self.band2]["href"] for feature in features]
-            if self.band2
-            else [None] * len(features)
+            [sign(feature["assets"][self.band2]["href"]) for feature in features]
+            if self.band2 else [None] * len(features)
         )
         return band1_urls, band2_urls
 
     def fetch_process_custom_band(self, band1_url, band2_url):
         try:
+            # print(f"Fetching Band1 URL: {band1_url}")
             with rasterio.open(band1_url) as band1_cog:
                 min_x, min_y, max_x, max_y = self._transform_bbox(band1_cog.crs)
-                band1_window = self._calculate_window(
-                    band1_cog, min_x, min_y, max_x, max_y
-                )
+                band1_window = self._calculate_window(band1_cog, min_x, min_y, max_x, max_y)
                 if self._is_window_out_of_bounds(band1_window):
-                    return None, None, None
-
+                    return None, None, None, None
                 band1 = band1_cog.read(window=band1_window).astype(float)
 
                 if band2_url:
+                    # print(f"Fetching Band2 URL: {band2_url}")
                     with rasterio.open(band2_url) as band2_cog:
-                        min_x, min_y, max_x, max_y = self._transform_bbox(band2_cog.crs)
-                        band2_window = self._calculate_window(
-                            band2_cog, min_x, min_y, max_x, max_y
-                        )
+                        band2_window = self._calculate_window(band2_cog, min_x, min_y, max_x, max_y)
                         if self._is_window_out_of_bounds(band2_window):
-                            return None, None, None
+                            return None, None, None, None
                         band2 = band2_cog.read(window=band2_window).astype(float)
-                        result = eval(self.formula)
+                        result = eval(self.formula, {}, {"band1": band1, "band2": band2})
[…]
