+++ src/vcube/extractors/landsat89_extractor.py
@@ -4,8 +4,11 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 
 import numpy as np
 import rasterio
+from planetary_computer import sign
+from pyproj import Transformer
 from rasterio.enums import Resampling
 from rasterio.warp import reproject
+from rasterio.windows import from_bounds
 from tqdm import tqdm
 
 from ..utils.common import (
@@ -15,23 +18,34 @@ from ..utils.common import (
     zip_files,
 )
 from ..utils.landsat89_utils import remove_overlapping_landsat_tiles
-from .extractor_common import ExtractorCommon
 
-# ✅ Landsat 8/9 supported bands
 VALID_BANDS = {
-    "coastal": "Coastal/Aerosol Band (Band 1)",
-    "blue": "Blue Band (Band 2)",
-    "green": "Green Band (Band 3)",
-    "red": "Red Band (Band 4)",
-    "nir": "Near Infrared Band (Band 5)",
-    "swir1": "Shortwave Infrared Band 1 (Band 6)",
-    "swir2": "Shortwave Infrared Band 2 (Band 7)",
-    "pan": "Panchromatic Band (Band 8)",
-    "cirrus": "Cirrus Band (Band 9)",
-    "lwir1": "Thermal Infrared Band 10",
-    "lwir2": "Thermal Infrared Band 11",
+    "coastal": "Coastal/Aerosol - 30m",
+    "blue": "Blue - 30m",
+    "green": "Green - 30m",
+    "red": "Red - 30m",
+    "nir08": "Near Infrared - 30m",
+    "swir16": "Shortwave Infrared 1 - 30m",
+    "swir22": "Shortwave Infrared 2 - 30m",
+    "lwir11": "Thermal Infrared (Band 10) - 30m",
+    "lwir12": "Thermal Infrared (Band 11) - 30m",
+
+    "qa_pixel": "Pixel QA - 30m",
+    "qa_radsat": "Radiometric Saturation QA - 30m",
+    "qa_aerosol": "Aerosol QA - 30m",
+
+    "st_atran": "Atmospheric Transmittance - 30m",
+    "st_cdist": "Cloud Distance - 30m",
+    "st_drad": "Downwelling Radiance - 30m",
+    "st_emis": "Emissivity - 30m",
+    "st_emsd": "Emissivity Standard Deviation - 30m",
+    "st_qa": "Surface Temperature QA - 30m",
+    "st_trad": "Thermal Radiance - 30m",
+    "st_urad": "Upwelling Radiance - 30m"
 }
-class ExtractProcessor(ExtractorCommon):
+
+
+class ExtractProcessor:
     def __init__(
         self,
         bbox,
@@ -57,7 +71,6 @@ class ExtractProcessor(ExtractorCommon):
         self.crs = None
         self.transform = None
         self.use_smart_filter = smart_filter
-
         self._validate_bands_list()
 
     def _validate_bands_list(self):
@@ -65,19 +78,38 @@ class ExtractProcessor(ExtractorCommon):
         if invalid_bands:
             raise ValueError(
                 f"Invalid band names: {', '.join(invalid_bands)}. "
-                f"Valid bands: {', '.join(VALID_BANDS.keys())}"
+                f"Band names should be one of: {', '.join(VALID_BANDS.keys())}"
             )
+
+    def _transform_bbox(self, crs):
+        transformer = Transformer.from_crs("epsg:4326", crs, always_xy=True)
+        min_x, min_y = transformer.transform(self.bbox[0], self.bbox[1])
+        max_x, max_y = transformer.transform(self.bbox[2], self.bbox[3])
+        return min_x, min_y, max_x, max_y
+
[…]
