### 2025-05-10 636f414
download:band list dropdown and satellite switch functionalities
