+++ src/vcube/engines/sentinel2_engine.py
@@ -1,526 +1,0 @@
-import os
-import sys
-from concurrent.futures import ThreadPoolExecutor, as_completed
-
-import matplotlib
-import matplotlib.pyplot as plt
-import numpy as np
-import rasterio
-from PIL import Image
-from pyproj import Transformer
-from rasterio.windows import from_bounds
-from tqdm import tqdm
-
-# ✅ Correct imports
-from ..utils.common import (
-    filter_intersected_features,
-    search_stac_api,
-    smart_filter_images,
-    zip_files,
-)
-from ..utils.sentinel2_utils import (
-    remove_overlapping_sentinel2_tiles,
-)
-
-matplotlib.use("Agg")
-
-
-class VCubeProcessor:
-    """
-    Processor for virtual computation cubes.
-    """
-
-    def __init__(
-        self,
-        bbox,
-        start_date,
-        end_date,
-        cloud_cover,
-        formula,
-        band1,
-        band2,
-        operation,
-        timeseries,
-        output_dir,
-        log_file=sys.stdout,
-        cmap="RdYlGn",
-        workers=1,
-        smart_filter=True,
-    ):
-        self.bbox = bbox
-        self.start_date = start_date
-        self.end_date = end_date
-        self.cloud_cover = cloud_cover
-        self.formula = formula or "band1"
-        self.band1 = band1
-        self.band2 = band2
-        self.operation = operation
-        self.timeseries = timeseries
-        self.output_dir = output_dir
-        self.log_file = log_file
-        self.cmap = cmap
-        self.workers = workers
-        self.result_list = []
-        self.dates = []
-        self.crs = None
-        self.transform = None
-        self.intermediate_images = []
-        self.intermediate_images_with_text = []
-        self.use_smart_filter = smart_filter
-    def fetch_process_custom_band(self, band1_url, band2_url):
-        try:
-            with rasterio.open(band1_url) as band1_cog:
-                min_x, min_y, max_x, max_y = self._transform_bbox(band1_cog.crs)
-                band1_window = self._calculate_window(
-                    band1_cog, min_x, min_y, max_x, max_y
-                )
-                if self._is_window_out_of_bounds(band1_window):
-                    return None, None, None
-
-                band1 = band1_cog.read(window=band1_window).astype(float)
-
-                if band2_url:
-                    with rasterio.open(band2_url) as band2_cog:
-                        min_x, min_y, max_x, max_y = self._transform_bbox(band2_cog.crs)
-                        band2_window = self._calculate_window(
-                            band2_cog, min_x, min_y, max_x, max_y
-                        )
-                        if self._is_window_out_of_bounds(band2_window):
-                            return None, None, None
-                        band2 = band2_cog.read(window=band2_window).astype(float)
-                        result = eval(self.formula)
-                else:
-                    result = eval(self.formula) if band1.shape[0] == 1 else band1
-
-            return (
-                result,
-                band1_cog.crs,
-                band1_cog.window_transform(band1_window),
-                band1_url,
-            )
-        except Exception as e:
-            print(f"Error fetching image: {e}")
-            return None, None, None
-
-    def _transform_bbox(self, crs):
-        transformer = Transformer.from_crs("epsg:4326", crs, always_xy=True)
-        min_x, min_y = transformer.transform(self.bbox[0], self.bbox[1])
-        max_x, max_y = transformer.transform(self.bbox[2], self.bbox[3])
-        return min_x, min_y, max_x, max_y
-
-    def _calculate_window(self, cog, min_x, min_y, max_x, max_y):
-        return from_bounds(min_x, min_y, max_x, max_y, cog.transform)
-
-    def _is_window_out_of_bounds(self, window):
-        return (
[…]
