+++ src/vcube/engines/sentinel2_engine1.py
@@ -9,7 +9,6 @@ import rasterio
 from tqdm import tqdm
 from PIL import Image
 
-# ✅ Correct imports
 from ..utils.common import (
     filter_intersected_features,
     search_stac_api,
