### 2025-05-10 3343a65
fix(bandversion): fix inconsistent band version
