### 2025-05-10 b23c900
smart filters working - corrected a typo
