### 2025-05-10 d5686fd
smart filter update finishing touch
