### 2025-06-04 #8078 @rwegener2
I apologize @rvg296, I mistyped. Generate your checklist with the following command:

You can read the more detailed set of starting instructions in [my first comment on this ticket](https://github.com/openjournals/joss-reviews/issues/8078#issuecomment-2815417794).
