### 2025-06-20 #8078 @rvg296
@kshitijrajsharma  The paper is well-articulated and explains how it handles and processes big earth data compared to other options.
As the other reviewer mentioned, it would be really helpful if you could include some quantitative and empiricial benchmarks in terms of computational efficiency, memory efficiency and processing handling. This would help the user to understand the efficacies involved in your approach. In addition, please try to include the fundamental comparison metrics with other big geospatial data frameworks, thereby reinforcing the uniqueness of the contributions.
