### 2025-08-22 02437e2
refactor(package): change package naming to avoid uppercase
BREAKING CHANGE:
