### 2025-08-22 ee1ff5d
bump: version 0.7.0 → 1.0.0
