+++ pyproject.toml
@@ -2,7 +2,7 @@
 name = "VirtuGhan"
 version = "0.7.0"
 description = "Virtual Computation Cube"
-authors = ["kshitijrajsharma <skshitizraj@gmail.com>"]
+authors = ["kshitijrajsharma <skshitizraj@gmail.com>", "geomatupen <upendra.oli59@gmail.com>"]
 readme = "README.md"
 packages = [
     { include = "vcube", from = "src" }
