### 2025-11-17 #8078 @kshitijrajsharma
Status : Draft

**Performance Comparison : Traditional vs Virtughan cloudnative approach**

Area of interest : Fewa Lake , Pokhara, Nepal
Theme : Mean NDWI analysis of fewa lake from 2025-10-17 to 2025-11-17 , Cloud cover : 30 %
BBOX: 83.91082763671876,28.18257204356296,84.00352478027345,28.23642498426942
Zoom Level: 14

**Performance using Virtughan :**

```log
Starting processing...

Engine starting...

Searching STAC .....

Total scenes found: 13

Scenes covering input area: 7

Scenes after removing overlaps: 7

Computing Band Calculation:   0%|          | 0/7 [00:00<?, ?it/s]

Computing Band Calculation:  14%|█▍        | 1/7 [00:24<02:28, 24.74s/it]

Computing Band Calculation:  29%|██▊       | 2/7 [00:28<01:00, 12.20s/it]

Computing Band Calculation:  43%|████▎     | 3/7 [00:30<00:30,  7.70s/it]

Computing Band Calculation:  57%|█████▋    | 4/7 [00:46<00:32, 10.93s/it]

Computing Band Calculation:  71%|███████▏  | 5/7 [00:49<00:16,  8.23s/it]

Computing Band Calculation:  86%|████████▌ | 6/7 [00:54<00:06,  6.91s/it]

Computing Band Calculation: 100%|██████████| 7/7 [00:56<00:00,  5.53s/it]

Computing Band Calculation: 100%|██████████| 7/7 [00:56<00:00,  8.12s/it]

Aggregating results...

Saving aggregated result with colormap...

Creating GIF and zipping TIFF files...

Saved timeseries GIF to static/export/20251117092401_a96f1d/output.gif

Saved intermediate images ZIP to static/export/20251117092401_a96f1d/tiff_files.zip

Processing completed. Results saved in static/export/20251117092401_a96f1d
```

Total time taken for band computation for all 7 Images found from Sentinel-2 , `00:56`  which is  8.12s per image in that area of interest !

**Traditional approach :**

TBD
