### 2026-01-07 #8078 @rvg296
Yes , some of the figures appear blurry in the PDF once I downloaded.> Thanks for that feedback @rvg296! Would you be able to provide a bit more detail for the author about what you mean by “improve the quality of figures”? Are you referring to the red spell check underline and the couple of grey-outlined boxes in Figure 1?
