### 2026-03-23 #8078 @rwegener2
Hi @kshitijrajsharma, just checking in. We have lots of submissions in the JOSS backlog awaiting review, so it would be helpful to those authors and to JOSS if you could respond to reviewer feedback and keep this paper submission moving along.

Are you able to respond to the latest round of reviewer comments by April 10th?
