### 2026-04-10 #8078 @geomatupen
Hi @rvg296 @rwegener2 , apologies for the delay. Here are the updated parts of the paper based on the feedback.
- Clarified that virtughan is conceptually similar to STAC/COG tile services such as TiTiler.
- Reframed virtughan as complementary to EO data cubes, not a replacement.
- Replaced “real-time” and “scalable” wording with more precise “on-demand” / “lightweight” language.
- Added implementation details on temporal aggregation, band resampling, and the lack of radiometric harmonization.
- Clarified that caching is local in-memory and maintained per process.
- Reframed the QGIS comparison as a workflow-level comparison rather than a direct computational benchmark.
- Added hardware and bandwidth specifications and clarified what timing components were excluded.
- Updated future work to mention the already-developed QGIS plugin.
- Improved wording around figures and benchmarking expectations.
- Added better quality image for workflow diagram and updated comparison diagram.
