### 2026-04-27 #8078 @rwegener2
Thanks @rvg296 and @jabhalla for your reviews! JOSS wouldn’t exist without you!

@kshitijrajsharma at this point I have a series of checks that I will go through as editor, before passing this submission on to the Track editor for final review.
