### 2026-04-27 #8078 @rwegener2
Hi @kshitijrajsharma,

Thanks for a well written paper! Just a few pieces of feedback:
- line 8: I’m confused by “endpoints COGEO (2024)”. That reference seems just to point to the cloud optimized geotiff landing page, but you’ve already mentioned cog in your list of tools. Please clarify or remove.
- line 39: Downloading does not need to be capitalized
- line 56-58: To clarify, where are you downloading the image tiles from? Is it Copernicus? Please clarify the source, and specify if the tiles were provided in cog format.
- line 80: captured at different time -> times
- Figure 1 caption: interesting figure! Please add a full caption that describes the table, the columns, and the relevance of row coloring. What is “SN” and “Divide time”? Is divide time how much faster the processing was? [Speedup](https://en.wikipedia.org/wiki/Speedup) is a standard metric you could use here, if you wish. As a suggestion, I would remove the Remark column and just mention the remark in the figure caption, but that is up to you.
- Figure 2: please add a brief caption. It doesn’t have to be long, but please convert from a hyphenated phrase to a sentence.
- line 119: Is it true that virtughan only works for some datasets? Please specify earlier in the paper which datasets or data providers it is compatible with.
- references: please provide dois for the two papers that have them (Giuliani et al and Sudmanns et al)

Let me know if you have any questions! After updating the paper you can complete the actions in the "Additional Author Tasks After Review is Complete” section of my checklist above.
