### 2026-04-28 #8078 @kshitijrajsharma
Thanks for the review @rwegener2
I have incorporated the feedback in the commit and referenced here !

>     * line 8: I’m confused by “endpoints COGEO (2024)”. That reference seems just to point to the cloud optimized geotiff landing page, but you’ve already mentioned cog in your list of tools. Please clarify or remove.

Removed. The cogeo citation duplicated cog and was redundant with the COG mention in the tags list. We dropped both the citation in the Summary and the duplicate entry in paper.bib.

>     * line 39: Downloading does not need to be capitalized

Fixed

>     * line 56-58: To clarify, where are you downloading the image tiles from? Is it Copernicus? Please clarify the source, and specify if the tiles were provided in cog format.

I have added clarification ;  The current implementation uses two STAC providers ( AWS Open Data and microsoft planetry computer ) , both of which expose Cloud-Optimized GeoTIFFs:

>     * line 80: captured at different time -> times
Fixed
>     * Figure 1 caption: interesting figure! Please add a full caption that describes the table, the columns, and the relevance of row coloring. What is “SN” and “Divide time”? Is divide time how much faster the processing was? [Speedup](https://en.wikipedia.org/wiki/Speedup) is a standard metric you could use here, if you wish. As a suggestion, I would remove the Remark column and just mention the remark in the figure caption, but that is up to you.
>
>     * Figure 2: please add a brief caption. It doesn’t have to be long, but please convert from a hyphenated phrase to a sentence.

I removed the colored chart and converted into the Markdown Table 1 in the Performance Comparison section, added speedup curve

>     * line 119: Is it true that virtughan only works for some datasets? Please specify earlier in the paper which datasets or data providers it is compatible with.

Yes ; currently two collections, and this is now stated upfront in the Summary (and elaborated in the Implementation section):

>     * references: please provide dois for the two papers that have them (Giuliani et al and Sudmanns et al)

Okay

> Let me know if you have any questions! After updating the paper you can complete the actions in the "Additional Author Tasks After Review is Complete” section of my checklist above.
