### 2026-05-01 #8078 @rwegener2
Looks good, thanks @kshitijrajsharma!

You can complete the tasks in section 2 of [my comment above](https://github.com/openjournals/joss-reviews/issues/8078#issuecomment-4328215187). After that I will submit the article to the track editor for the final check.
